Refractor and improve naming conventions (#761)
