+++ src/test/java/SimulatorParserTest.java
@@ -1,21 +1,9 @@
-import com.appium.ios.SimManager;
-import com.appium.manager.AppiumDriverManager;
-import org.openqa.selenium.remote.DesiredCapabilities;
 import org.testng.annotations.Test;
 
-/**
- * Created by saikrisv on 14/08/17.
- */
 public class SimulatorParserTest {
 
-    SimManager simManager;
-    AppiumDriverManager appiumDriverManager = new AppiumDriverManager();
-
-    public SimulatorParserTest() throws Exception {
-    }
-
     @Test
-    public void parseSimulatorJson() throws Exception {
+    public void parseSimulatorJson() {
         String version = "9.1.1";
         Float aFloat = Float.valueOf(version.substring(0, version.length() - 2));
         System.out.println(aFloat);
