+++ src/test/resources/XMLSuiteForMethodLevelDistribution.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite thread-count="4" verbose="2" name="TestNG Forum" parallel="methods" data-provider-thread-count="4">
+    <listeners>
+        <listener class-name="com.appium.manager.AppiumParallelMethodTestListener"/>
+        <listener class-name="com.appium.utils.RetryListener"/>
+        <listener class-name="com.appium.manager.AppiumParallelMethodTestListener"/>
+        <listener class-name="com.appium.utils.RetryListener"/>
+    </listeners>
+    <test thread-count="4" verbose="2" name="TestNG Test" parallel="methods">
+        <parameter name="device" value=""/>
+        <classes>
+            <class name="com.appium.executor.OtherTests1"/>
+            <class name="com.appium.executor.OtherTests"/>
+            <class name="com.appium.executor.ATDExecutorTest"/>
+        </classes>
+    </test> <!-- TestNG Test -->
+</suite> <!-- TestNG Forum -->
