+++ src/test/resources/XMLSuiteForParallelRunner.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite thread-count="4" verbose="2" name="TestNG Forum" parallel="tests" data-provider-thread-count="4">
+    <listeners>
+        <listener class-name="com.appium.manager.AppiumParallelTestListener"/>
+        <listener class-name="com.appium.utils.RetryListener"/>
+    </listeners>
+    <test thread-count="4" verbose="2" name="TestNG Test-0" parallel="tests" preserve-order="false">
+        <parameter name="hostName" value="10.10.10.10"/>
+        <classes>
+            <class name="com.appium.executor.OtherTests1"/>
+            <class name="com.appium.executor.OtherTests"/>
+            <class name="com.appium.executor.ATDExecutorTest"/>
+        </classes>
+    </test> <!-- TestNG Test-0 -->
+    <test thread-count="4" verbose="2" name="TestNG Test-1" parallel="tests" preserve-order="false">
+        <parameter name="hostName" value="10.10.10.10"/>
+        <classes>
+            <class name="com.appium.executor.OtherTests1"/>
+            <class name="com.appium.executor.OtherTests"/>
+            <class name="com.appium.executor.ATDExecutorTest"/>
+        </classes>
+    </test> <!-- TestNG Test-1 -->
+    <test thread-count="4" verbose="2" name="TestNG Test-2" parallel="tests" preserve-order="false">
+        <parameter name="hostName" value="10.10.10.10"/>
+        <classes>
+            <class name="com.appium.executor.OtherTests1"/>
+            <class name="com.appium.executor.OtherTests"/>
+            <class name="com.appium.executor.ATDExecutorTest"/>
+        </classes>
+    </test> <!-- TestNG Test-2 -->
+    <test thread-count="4" verbose="2" name="TestNG Test-3" parallel="tests" preserve-order="false">
+        <parameter name="hostName" value="10.10.10.10"/>
+        <classes>
+            <class name="com.appium.executor.OtherTests1"/>
+            <class name="com.appium.executor.OtherTests"/>
+            <class name="com.appium.executor.ATDExecutorTest"/>
+        </classes>
+    </test> <!-- TestNG Test-3 -->
+</suite> <!-- TestNG Forum -->
